Change the locus_tag prefix from OU3MT to ABT39_MT
On 2015-09-17 Shelby Bidwell wrote...
We noticed that you used the locus_tag prefix "OU3MT" in your submission file.
"OU3" is already assigned to another BioSample (SAMN02223065). The locus_tag
prefix that is assigned to SAMN01120252 is "ABT39". In order to distinguish
this from nuclear genome that is associated with this BioSample, you can use
ABT39_MTgene1, ABT39_MTgene2, and so on.