Add pg29mt-scaffolds.pep for RNA editing
Add translations for genes with RNA editing.

On 2015-09-17 Shelby Bidwell wrote...
You will need to remove the "transl_except" from these CDS, and instead,
import the edited translation with the "exception".

On 2015-09-22 Shelby Bidwell wrote...
What you'll need to do is to put the translation(s) in a .pep file that has the
same basename as the .fsa file, and the identifier of each translation must
match the protein_id in the .tbl file, and put the .pep file in the same
directory as the .fsa and .tbl files. See the "Protein sequence format (.pep) "
section of http://www.ncbi.nlm.nih.gov/genbank/tbl2asn2/

Here is an example of what your .pep file should look like:

>gnl|BCGSC|ABT39_MTgene3056
MAQRQIWFGIPLFPVLVMFFISRPAETNR…

This .pep file will substitute the automatically translated products of the CDS
features with the provided protein sequences, and is only needed in unusual
cases. Your RNA editing exception would stay in your .tbl file as it was. Make
sure to remove the transl_except.